Added ProbeSchedule class
The ProbeSchedule class, if not left empty, allows the user to schedule the x- and y-offsets of the probes (at the DM) as well as the probe intensity at each WFSC iteration.

+++ lib/wfsc/falco_gen_pairwise_probe.m
@@ -52,8 +52,8 @@ Nact = dm.Nact;
 NactPerBeam = mp.P2.D / dm.dm_spacing;
 
 % Coordinates in actuator space
-xs = (-(Nact-1)/2:(Nact-1)/2)/Nact - round(mp.est.probe.xOffset)/Nact;
-ys = (-(Nact-1)/2:(Nact-1)/2)/Nact - round(mp.est.probe.yOffset)/Nact;
+xs = (-(Nact-1)/2:(Nact-1)/2)/Nact - double(round(mp.est.probe.xOffset))/Nact;
+ys = (-(Nact-1)/2:(Nact-1)/2)/Nact - double(round(mp.est.probe.yOffset))/Nact;
 [XS, YS] = meshgrid(xs, ys);
 
 % Convert units from lambda/D to actuators
